Added new name finding mechanism for sound names.
Actual build of dialog editor. Zlib Byte changed to ByteZ for escape ambiguous symbol error in compilation.

git-svn-id: https://cvet.by:8443/svn/FOnlineSource@77 c2c6a35e-ad2b-4e1f-bb64-97b989be360b

Former-commit-id: a7f4577f5582ed403d62ada63559b81836f630be

+++ Source/SoundManager.h
@@ -95,7 +95,7 @@ public:
 
 	bool LoadSoundList(const char* lst_name, int path_type);
 	void PlaySound(const char* name);
-	void PlayAction(const char* body_type, BYTE anim1, BYTE anim2);
+	void PlayAction(const char* body_type, DWORD anim1, DWORD anim2);
 	void PlaySoundType(BYTE sound_type, BYTE sound_type_ext, BYTE sound_id, BYTE sound_id_ext);
 	void PlayMusic(const char* fname, DWORD pos = 0, DWORD repeat = MUSIC_PLAY_PAUSE);
 	void StopMusic();
